+++ src/modules/iframe/components/chips/ViewLargerMapChip.js
